+++ MochiKit/Signal.js
@@ -8,7 +8,7 @@ See <http://mochikit.com/> for documentation, downloads, license, etc.
 
 ***/
 
-MochiKit.Base._module('Signal', '1.5', ['Base', 'DOM', 'Style']);
+MochiKit.Base._module('Signal', '1.5', ['Base', 'DOM']);
 
 MochiKit.Signal._observers = [];
 
@@ -270,13 +270,13 @@ MochiKit.Base.update(MochiKit.Signal.Event.prototype, {
             this.type().indexOf('click') != -1 ||
             this.type() == 'contextmenu')) {
 
-            m.client = new MochiKit.Style.Coordinates(0, 0);
+            m.client = { x: 0, y: 0 };
             if (e.clientX || e.clientY) {
                 m.client.x = (!e.clientX || e.clientX < 0) ? 0 : e.clientX;
                 m.client.y = (!e.clientY || e.clientY < 0) ? 0 : e.clientY;
             }
 
-            m.page = new MochiKit.Style.Coordinates(0, 0);
+            m.page = { x: 0, y: 0 };
             if (e.pageX || e.pageY) {
                 m.page.x = (!e.pageX || e.pageX < 0) ? 0 : e.pageX;
                 m.page.y = (!e.pageY || e.pageY < 0) ? 0 : e.pageY;
@@ -338,7 +338,7 @@ MochiKit.Base.update(MochiKit.Signal.Event.prototype, {
                 }
             }
             if (this.type() == 'mousewheel') {
-                m.wheel = new MochiKit.Style.Coordinates(0, 0);
+                m.wheel = { x: 0, y: 0 };
                 if (e.wheelDeltaX || e.wheelDeltaY) {
                     m.wheel.x = e.wheelDeltaX / -40 || 0;
                     m.wheel.y = e.wheelDeltaY / -40 || 0;