+++ doc/rst/MochiKit/VersionHistory.rst
@@ -36,6 +36,7 @@
 - Added state() method for MochiKit.Async.Deferred. Also clarified docs on
   available properties.
 - Added MochiKit.DOM.IFRAME, LINK and SCRIPT functions.
+- Removed MochiKit.Style dependency from MochiKit.Signal.
 
 2009-XX-YY      v1.4.3 (bug fix release)